Refactor __init__.py structure and identify internal import issues
This commit addresses the improper packaging of the NDML library by restructuring the __init__.py files.

Key changes:
- I broke down the monolithic root __init__.py file.
- I moved code specific to 'core', 'integration', 'deployment', 'consensus', and 'utils' submodules into their respective __init__.py files.
- I created the 'consensus' and 'utils' directories and their __init__.py files as they were missing from my initial analysis.
- I cleaned the root __init__.py to only contain top-level package information and imports from its submodules.

Further work identified:
During verification, I discovered that many .py files within the submodules use absolute import paths (e.g., 'from core.x') instead of relative paths (e.g., 'from ..core.x' or 'from .core.x'). This prevents the package from loading correctly. My next steps will involve correcting these internal import paths and then re-verifying the package functionality.
