+++ consensus/__init__.py
@@ -1,0 +1,14 @@
+# consensus/__init__.py - Consensus components package
+"""
+NDML Consensus Components
+
+Distributed consensus mechanisms for coordinating memory updates
+across multiple nodes using neuromorphic principles.
+"""
+
+# Note: consensus/neuromorphic.py is referenced but not yet implemented
+# This would contain the NeuromorphicConsensusLayer class
+
+__all__ = [
+    # Will be populated when consensus components are implemented
+]
