+++ deployment/__init__.py
@@ -1,0 +1,14 @@
+# deployment/__init__.py - Deployment components package
+"""
+NDML Deployment Components
+
+Tools and configurations for deploying NDML in production environments,
+including Kubernetes manifests and consensus nodes.
+"""
+
+from .consensus_node import ConsensusNode, create_consensus_node
+
+__all__ = [
+    'ConsensusNode',
+    'create_consensus_node',
+]
